# openapi.yaml
openapi: 3.1.0
info:
  title: pogolo API
  description: API specification for pogolo
  version: 1.0.0
  license:
    name: GNU General Public License version 3
    url: https://opensource.org/license/gpl-3-0
servers:
  - url: http://{server}:{port}
    variables:
      server:
        default: "127.0.0.1"
      port:
        default: "5662"
components:
  schemas:
...
    GopherInfo:
      type: object
      required:
        - address
        - averageShareTime
        - bestDifficulty
        - extranonce1
        - hashrate
        - nickname
        - protocolVersion
        - sharesAccepted
        - sharesRejected
        - targetDifficulty
        - uptime
        - userAgent
      properties:
        address:
          type: string
          description: On-chain address the miner is mining to
        averageShareTime:
          type: number
          description: float64 averaged share time, in seconds
        bestDifficulty:
          type: number
        extranonce1:
          type: string
        hashrate:
          type: number
          description: In MH/s
        nickname:
          type: string
        protocolVersion:
          type: number
          description: either 1 for SV1, or 2 for SV2
        sharesAccepted:
          type: integer
        sharesRejected:
          type: integer
        targetDifficulty:
          type: number
        uptime:
          type: integer
          description: In seconds
        userAgent:
          type: string
...
paths:
...
  /api/v1/gophers/{idOrNickname}:
    get:
      summary: Get snapshot of a gophers stats
      operationId: getGopherInfo
      parameters:
        - name: idOrNickname
          in: path
          required: true
          description: The gophers extranonce1 or nickname
          schema:
            type: string

      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GopherInfo"
...